Fix checkstyle violations
