+++ data/PloopyDatabase.txt
@@ -1,7 +1,2 @@
 T_0_homework_0
 D_0_assignment 1_12/08/2022 1800_0
-E_0_dinner_13/09/2022 2000_0
-T_0_cs2100 assignment_0
-E_0_tutorial_14/09/2022 0800_0
-E_0_tut_12/12/2022 0800_0
-E_0_tut_12/12/2022 0800_0
