+++ src/main/java/ploopy/task/Event.java
@@ -1,13 +1,30 @@
 package ploopy.task;
 
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+/**
+ * Represents an event task that has a date as well as
+ * a name.
+ */
 public class Event extends Task {
+    /** Input string to Date formatter */
+    private static final DateTimeFormatter INPUT_FORMATTER = DateTimeFormatter.ofPattern("dd/MM/yyyy kkmm");
+    /** Date to output String formatter */
+    private static final DateTimeFormatter OUTPUT_FORMATTER = DateTimeFormatter.ofPattern("MMM dd yyyy kkmm");
+    /** Date of task (If applicable) */
+    private LocalDateTime dateTime;
+    /** String format of date */
+    private String dateStringForm;
 
     /**
      * {@inheritDoc}
      */
     public Event(String name, String date) {
-        super(name, date);
+        super(name);
         type = "E";
+        dateStringForm = date;
+        this.dateTime = LocalDateTime.parse(date, INPUT_FORMATTER);
     }
 
     @Override
@@ -20,6 +37,11 @@ public class Event extends Task {
      */
     @Override
     public String getDate() {
-        return String.format("(at: %s)", super.getDate());
+        return String.format("(at: %s)", OUTPUT_FORMATTER.format(dateTime));
+    }
+
+    @Override
+    public String getDateForFileWrite() {
+        return dateStringForm;
     }
 }
